// types.ts
export type TransportMode =
  | "car"
  | "walking"
  | "bicycle"
  | "bus"
  | "train"
  | "subway"
  | "running"
  | "still";
// ...
export interface LocationData {
  latitude: number;
  longitude: number;
  altitude?: number;
  speed?: number;
  heading?: number;
  accuracy: number;
  timestamp: number;
  placeName?: string;
  placeType?: string;
}